Use internal state for RenameDialog component's form
We don't need to keep the form state across open/close of the rename
dialog.
By resetting the form on the opening of the dialog, we also make sure
the correct data is shown based on which favorite is selected.
Fixed also the payload sent in the PATCH request, as the description is
not required and can be empty.